+++ app/components/deploy/GitHubDeploy.client.tsx
@@ -145,6 +145,9 @@ export function useGitHubDeploy() {
         source: 'github',
       });
 
+      // Show success toast notification
+      toast.success(`🚀 GitHub deployment preparation completed successfully!`);
+
       return {
         success: true,
         files: fileContents,
