+++ app/components/deploy/GitLabDeploy.client.tsx
@@ -145,6 +145,9 @@ export function useGitLabDeploy() {
         source: 'gitlab',
       });
 
+      // Show success toast notification
+      toast.success(`🚀 GitLab deployment preparation completed successfully!`);
+
       return {
         success: true,
         files: fileContents,
